refactor(prompts): tighten review fixes + dedupe wrapper-WS plumbing
Follow-up to the prompts module audit. Five fixes in one coherent pass:

- **Promote wrapper-frame-sender to core/.** Approvals' delivery sweeper
  and prompts' forwarder both opened a one-shot WS, awaited open, sent
  one frame, drained, closed — character-for-character identical
  plumbing in two places. Move the factory + interface into
  `core/wrapper-frame-sender.ts` (peer of `redis-bus.ts`,
  `acp-client.ts`). Approvals re-exports its `WrapperFrameSender` port
  for type continuity. Prompts' `wrapper-prompt-forwarder.ts` shrinks
  from 87 lines (full WS dance + helpers) to 40 (just the
  frame-builder). Net −50 lines of duplication.

- **TRPCError on ownership failure.** `prompts.send` now throws
  `TRPCError({ code: "NOT_FOUND" })` instead of plain Error — was
  surfacing as 500, hides the real status from clients.

- **Tighten router validation.** Route `prompt` field through ACP SDK's
  `zContentBlock` schema instead of `z.array(z.unknown())`. Defense in
  depth; the wrapper revalidates downstream, but rejecting at the API
  edge fails loud + fast on garbage payloads. Adds `@agentclientprotocol/sdk`
  as a direct dep on `api-server-api` (zero transitive runtime deps).

- **Wire `clearLocalPromptIds` to session change.** Was dead code.
  Without it, switching session and back lost the user's just-sent
  message: cold-replay `user_message_chunk`s carry the same `_meta.promptId`
  the local skip-set has, so the projection filtered them out and the
  rebuilt conversation was missing the prompt.

- **Rename `_promptsStore` → `promptsStore`.** Underscore prefix
  conventionally signals "intentionally unused"; the binding is actively
  used twice.

Documents the wrapper-rejection trade-off (PROMPT_QUEUE_CAP exceeded /
session-not-found responses are silently dropped) inline as a known
shortcoming for future tightening.
