+++ packages/api-server-api/package.json
@@ -12,6 +12,7 @@
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
+    "@agentclientprotocol/sdk": "^0.17.1",
     "@trpc/server": "^11.16.0",
     "zod": "^4.3.6"
   },
